Monte Carlo predictive sampler for spline regression

Compute direct Monte Carlo samples from the posterior predictive distribution of a STAR spline regression model.

Usage

spline_star_exact(
  y,
  tau = NULL,
  transformation = "np",
  y_max = Inf,
  psi = 1000,
  method_sigma = "mle",
  approx_Fz = FALSE,
  approx_Fy = FALSE,
  nsave = 1000,
  compute_marg = TRUE
)

Arguments

y

n x 1 vector of observed counts

tau

n x 1 vector of observation points; if NULL, assume equally-spaced on [0,1]

transformation

transformation to use for the latent data; must be one of

"identity" (identity transformation)

"log" (log transformation)

"sqrt" (square root transformation)

"bnp" (Bayesian nonparametric transformation using the Bayesian bootstrap)

"np" (nonparametric transformation estimated from empirical CDF)

"pois" (transformation for moment-matched marginal Poisson CDF)

"neg-bin" (transformation for moment-matched marginal Negative Binomial CDF)

y_max

a fixed and known upper bound for all observations; default is Inf

psi

prior variance (1/smoothing parameter)

method_sigma

method to estimate the latent data standard deviation; must be one of

"mle" use the MLE from the STAR EM algorithm

"mmle" use the marginal MLE (Note: slower!)

approx_Fz

logical; in BNP transformation, apply a (fast and stable) normal approximation for the marginal CDF of the latent data

approx_Fy

logical; in BNP transformation, approximate the marginal CDF of y using the empirical CDF

nsave

number of Monte Carlo simulations

compute_marg

logical; if TRUE, compute and return the marginal likelihood

Value

a list with the following elements:

post_ytilde: nsave x n samples from the posterior predictive distribution at the observation points tau

marg_like: the marginal likelihood (if requested; otherwise NULL)

Details

STAR defines a count-valued probability model by (1) specifying a Gaussian model for continuous *latent* data and (2) connecting the latent data to the observed data via a *transformation and rounding* operation. Here, the continuous latent data model is a spline regression.

There are several options for the transformation. First, the transformation can belong to the *Box-Cox* family, which includes the known transformations 'identity', 'log', and 'sqrt'. Second, the transformation can be estimated (before model fitting) using the empirical distribution of the data y. Options in this case include the empirical cumulative distribution function (CDF), which is fully nonparametric ('np'), or the parametric alternatives based on Poisson ('pois') or Negative-Binomial ('neg-bin') distributions. For the parametric distributions, the parameters of the distribution are estimated using moments (means and variances) of y. The distribution-based transformations approximately preserve the mean and variance of the count data y on the latent data scale, which lends interpretability to the model parameters. Lastly, the transformation can be modeled using the Bayesian bootstrap ('bnp'), which is a Bayesian nonparametric model and incorporates the uncertainty about the transformation into posterior and predictive inference.

The Monte Carlo sampler produces direct, discrete, and joint draws from the posterior predictive distribution of the spline regression model at the observed tau points.

Examples

# Simulate some data:
n = 100
tau = seq(0,1, length.out = n)
y = round_floor(exp(1 + rnorm(n)/4 + poly(tau, 4)%*%rnorm(n=4, sd = 4:1)))

# Sample from the predictive distribution of a STAR spline model:
fit_star = STAR_spline(y = y, tau = tau)
#> Error in STAR_spline(y = y, tau = tau): could not find function "STAR_spline"
post_ytilde = fit_star$post_ytilde
#> Error in eval(expr, envir, enclos): object 'fit_star' not found

# Compute 90% prediction intervals:
pi_y = t(apply(post_ytilde, 2, quantile, c(0.05, .95)))
#> Error in apply(post_ytilde, 2, quantile, c(0.05, 0.95)): object 'post_ytilde' not found

# Plot the results: intervals, median, and smoothed mean
plot(tau, y, ylim = range(pi_y, y))
#> Error in plot.default(tau, y, ylim = range(pi_y, y)): object 'pi_y' not found
polygon(c(tau, rev(tau)),c(pi_y[,2], rev(pi_y[,1])),col='gray', border=NA)
#> Error in xy.coords(x, y, setLab = FALSE): object 'pi_y' not found
lines(tau, apply(post_ytilde, 2, median), lwd=5, col ='black')
#> Error in apply(post_ytilde, 2, median): object 'post_ytilde' not found
lines(tau, smooth.spline(tau, apply(post_ytilde, 2, mean))$y, lwd=5, col='blue')
#> Error in apply(post_ytilde, 2, mean): object 'post_ytilde' not found
lines(tau, y, type='p')
#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet